+++ src/main/java/com/epam/jwd/model/DBEntity.java
@@ -3,5 +3,5 @@ package com.epam.jwd.model;
 import java.io.Serializable;
 
 public interface DBEntity extends Serializable {
-    Long getAccountId();
+   int getAccountId();
 }
